feat(ios-refund,card-news,dormant-likes,deleted-females,reset-password): add v2 pages with BFF proxy
Extract legacy pages and create v2 variants with patchAdminAxios.
Standard conversions with server component routers gated on feature flags.
