+++ app/admin/ios-refund/page.tsx
@@ -1,346 +1,14 @@
-'use client';
+import { getRouteMode, isAdminShellV2Enabled } from '@/shared/feature-flags';
+import IOSRefundPageLegacy from './ios-refund-legacy';
+import IOSRefundPageV2 from './ios-refund-v2';
 
-import { useState, useEffect, useCallback } from 'react';
-import AdminService from '@/app/services/admin';
-import { AppleRefundItem, AppleRefundStatus, AppleRefundListParams } from '@/types/admin';
-import { LegacyPageAdapter } from '@/shared/ui/admin/legacy-page-adapter';
+export default async function IOSRefundPage() {
+  const shellV2 = await isAdminShellV2Enabled();
+  const mode = await getRouteMode('ios-refund');
 
-type FilterStatus = 'ALL' | AppleRefundStatus;
+  if (shellV2 && mode === 'v2') {
+    return <IOSRefundPageV2 />;
+  }
 
-function IOSRefundPageContent() {
-  const [items, setItems] = useState<AppleRefundItem[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [syncing, setSyncing] = useState(false);
-  const [page, setPage] = useState(1);
-  const [totalPages, setTotalPages] = useState(1);
-  const [totalCount, setTotalCount] = useState(0);
-  const [filterStatus, setFilterStatus] = useState<FilterStatus>('ALL');
-  const [searchTerm, setSearchTerm] = useState('');
-  const [startDate, setStartDate] = useState('');
-  const [endDate, setEndDate] = useState('');
-
-  const fetchData = useCallback(async () => {
-    try {
-      setLoading(true);
-      const params: AppleRefundListParams = {
-        page,
-        limit: 20,
-        ...(filterStatus !== 'ALL' && { status: filterStatus }),
-        ...(searchTerm && { searchTerm }),
-        ...(startDate && { startDate }),
-        ...(endDate && { endDate }),
-      };
-      const response = await AdminService.appleRefund.getList(params);
-      setItems(response.items);
-      setTotalPages(response.meta.totalPages);
-      setTotalCount(response.meta.totalCount);
-    } catch (error) {
-      console.error('iOS 환불 내역 조회 실패:', error);
-    } finally {
-      setLoading(false);
-    }
-  }, [page, filterStatus, searchTerm, startDate, endDate]);
-
-  useEffect(() => {
-    fetchData();
-  }, [fetchData]);
-
-  const handleSync = async () => {
-    try {
-      setSyncing(true);
-      await AdminService.appleRefund.syncRefundStatus();
-      await fetchData();
-      alert('환불 상태 동기화가 완료되었습니다.');
-    } catch (error) {
-      console.error('동기화 실패:', error);
-      alert('동기화에 실패했습니다.');
-    } finally {
-      setSyncing(false);
-    }
-  };
-
-  const handleSearch = () => {
-    setPage(1);
-    fetchData();
-  };
-
-  const handleReset = () => {
-    setFilterStatus('ALL');
-    setSearchTerm('');
-    setStartDate('');
-    setEndDate('');
-    setPage(1);
-  };
-
-  const getStatusBadge = (status: AppleRefundStatus) => {
-    switch (status) {
-      case AppleRefundStatus.REFUNDED:
-        return (
-          <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-red-100 text-red-800">
-            환불 완료
-          </span>
-        );
-      case AppleRefundStatus.NONE:
-      default:
-        return (
-          <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-green-100 text-green-800">
-            정상
-          </span>
-        );
-    }
-  };
-
-  const formatDate = (dateString: string | null) => {
-    if (!dateString) return '-';
-    return new Date(dateString).toLocaleString('ko-KR', {
-      year: 'numeric',
-      month: '2-digit',
-      day: '2-digit',
-      hour: '2-digit',
-      minute: '2-digit',
-    });
-  };
-
-  const formatCurrency = (amount: number, currency: string) => {
-    return new Intl.NumberFormat('ko-KR', {
-      style: 'currency',
-      currency: currency || 'KRW',
-    }).format(amount);
-  };
-
-  return (
-    <div className="min-h-screen bg-gray-50">
-      <div className="bg-white shadow-sm border-b border-gray-200">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-6">
-          <div className="flex justify-between items-center">
-            <div>
-              <h1 className="text-3xl font-bold text-gray-900">iOS 환불 관리</h1>
-              <p className="mt-2 text-sm text-gray-600">
-                Apple App Store 인앱 결제 환불 내역을 관리합니다
-              </p>
-            </div>
-            <div className="flex items-center space-x-3">
-              <button
-                onClick={handleSync}
-                disabled={syncing}
-                className={`px-4 py-2 rounded-lg text-white font-medium transition-colors ${
-                  syncing
-                    ? 'bg-gray-400 cursor-not-allowed'
-                    : 'bg-blue-600 hover:bg-blue-700'
-                }`}
-              >
-                {syncing ? '동기화 중...' : '환불 상태 동기화'}
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
-        <div className="space-y-6">
-          <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6">
-            <h2 className="text-lg font-semibold text-gray-900 mb-4">필터</h2>
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4">
-              <div>
-                <label className="block text-sm font-medium text-gray-700 mb-1">
-                  환불 상태
-                </label>
-                <select
-                  value={filterStatus}
-                  onChange={(e) => setFilterStatus(e.target.value as FilterStatus)}
-                  className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-purple-500 focus:border-transparent"
-                >
-                  <option value="ALL">전체</option>
-                  <option value={AppleRefundStatus.NONE}>정상</option>
-                  <option value={AppleRefundStatus.REFUNDED}>환불 완료</option>
-                </select>
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-gray-700 mb-1">
-                  사용자 검색
-                </label>
-                <input
-                  type="text"
-                  value={searchTerm}
-                  onChange={(e) => setSearchTerm(e.target.value)}
-                  placeholder="사용자 이름 또는 ID"
-                  className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-purple-500 focus:border-transparent"
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-gray-700 mb-1">
-                  시작일
-                </label>
-                <input
-                  type="date"
-                  value={startDate}
-                  onChange={(e) => setStartDate(e.target.value)}
-                  className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-purple-500 focus:border-transparent"
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-gray-700 mb-1">
-                  종료일
-                </label>
-                <input
-                  type="date"
-                  value={endDate}
-                  onChange={(e) => setEndDate(e.target.value)}
-                  className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-purple-500 focus:border-transparent"
-                />
-              </div>
-            </div>
-            <div className="flex justify-end gap-2 mt-4">
-              <button
-                onClick={handleReset}
-                className="px-4 py-2 text-gray-700 bg-gray-100 rounded-lg hover:bg-gray-200 transition-colors"
-              >
-                초기화
-              </button>
-              <button
-                onClick={handleSearch}
-                className="px-4 py-2 text-white bg-purple-600 rounded-lg hover:bg-purple-700 transition-colors"
-              >
-                검색
-              </button>
-            </div>
-          </div>
-
-          <div className="bg-white rounded-lg shadow-sm border border-gray-200">
-            <div className="px-6 py-4 border-b border-gray-200 flex justify-between items-center">
-              <h2 className="text-lg font-semibold text-gray-900">
-                환불 내역 <span className="text-gray-500 font-normal">({totalCount}건)</span>
-              </h2>
-            </div>
-
-            {loading ? (
-              <div className="flex justify-center items-center py-20">
-                <div className="w-8 h-8 border-4 border-purple-600 border-t-transparent rounded-full animate-spin" />
-              </div>
-            ) : items.length === 0 ? (
-              <div className="flex flex-col items-center justify-center py-20 text-gray-500">
-                <svg
-                  className="w-16 h-16 mb-4 text-gray-300"
-                  fill="none"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={1.5}
-                    d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"
-                  />
-                </svg>
-                <p>환불 내역이 없습니다</p>
-              </div>
-            ) : (
-              <div className="overflow-x-auto">
-                <table className="w-full">
-                  <thead className="bg-gray-50">
-                    <tr>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        사용자
-                      </th>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        상품 ID
-                      </th>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        거래 ID
-                      </th>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        금액
-                      </th>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        결제일
-                      </th>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        환불일
-                      </th>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        상태
-                      </th>
-                    </tr>
-                  </thead>
-                  <tbody className="bg-white divide-y divide-gray-200">
-                    {items.map((item) => (
-                      <tr key={item.id} className="hover:bg-gray-50">
-                        <td className="px-6 py-4 whitespace-nowrap">
-                          <div className="text-sm font-medium text-gray-900">
-                            {item.userName}
-                          </div>
-                          <div className="text-sm text-gray-500">{item.userId}</div>
-                        </td>
-                        <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">
-                          {item.productId}
-                        </td>
-                        <td className="px-6 py-4 whitespace-nowrap">
-                          <div className="text-sm text-gray-900">{item.transactionId}</div>
-                          <div className="text-xs text-gray-500">
-                            원본: {item.originalTransactionId}
-                          </div>
-                        </td>
-                        <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">
-                          {formatCurrency(item.amount, item.currency)}
-                        </td>
-                        <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                          {formatDate(item.purchaseDate)}
-                        </td>
-                        <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                          {formatDate(item.refundDate)}
-                        </td>
-                        <td className="px-6 py-4 whitespace-nowrap">
-                          {getStatusBadge(item.refundStatus)}
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-            )}
-
-            {totalPages > 1 && (
-              <div className="px-6 py-4 border-t border-gray-200 flex items-center justify-between">
-                <div className="text-sm text-gray-500">
-                  {page} / {totalPages} 페이지
-                </div>
-                <div className="flex gap-2">
-                  <button
-                    onClick={() => setPage((p) => Math.max(1, p - 1))}
-                    disabled={page === 1}
-                    className={`px-4 py-2 rounded-lg text-sm font-medium transition-colors ${
-                      page === 1
-                        ? 'bg-gray-100 text-gray-400 cursor-not-allowed'
-                        : 'bg-gray-100 text-gray-700 hover:bg-gray-200'
-                    }`}
-                  >
-                    이전
-                  </button>
-                  <button
-                    onClick={() => setPage((p) => Math.min(totalPages, p + 1))}
-                    disabled={page === totalPages}
-                    className={`px-4 py-2 rounded-lg text-sm font-medium transition-colors ${
-                      page === totalPages
-                        ? 'bg-gray-100 text-gray-400 cursor-not-allowed'
-                        : 'bg-gray-100 text-gray-700 hover:bg-gray-200'
-                    }`}
-                  >
-                    다음
-                  </button>
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-export default function IOSRefundPage() {
-  return (
-    <LegacyPageAdapter>
-      <IOSRefundPageContent />
-    </LegacyPageAdapter>
-  );
+  return <IOSRefundPageLegacy />;
 }
